factor out hacker news integration into own module
